ingestion/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
aiokafka==0.11.0
aiohttp==3.10.5
asyncpg==0.29.0
redis[hiredis]==5.1.1
clickhouse-connect==0.8.0
ua-parser==0.18.0
geoip2==4.8.0
orjson==3.10.7
httpx==0.27.2
structlog==24.4.0
argon2-cffi==23.1.0
